More refactoring. Impl of config for channel, setup, filter...
AD411x_Device is now an abstract base class with generic impl.
AD4116_Device is an object class that represents an AD4116 ADC.
AD4116_Device contains only impl specific to AD4116.
AD4116 is a namespace with definitions specific to that device.
AD411x is a namespace with definitions shared across the AD411x family.